requirements.txt
guardrails-ai
openai
python-dotenv
requests 
uuid
json